# BookNest/backend/public/openapi.yaml
openapi: 3.0.3
info:
  title: BookNest API
  description: >
    Documentation for the two BookNest backends. The Rails application
    (BookNest/backend, :8080) is the public JSON API for the catalog, cart,
    orders and the Ruby playground used by the React storefront
    (frontend-shop) — plus a separate Spring Boot auth-service (:9000) that
    handles registration/login/Google OAuth and issues JWTs. The Rails
    backend does not store its own credentials; it only verifies the
    signature of the JWT issued by auth-service, carried in the
    `booknest_jwt` cookie.
  version: "2.0.0"
servers:
  - url: http://localhost:8080
    description: Rails backend (BookNest JSON API + docker compose / rails s)
  - url: http://localhost:9000
    description: Spring Boot auth-service (docker compose)
...
paths:
...
  /api/auth/register:
    post:
      summary: Register a new user
      tags: [Auth (auth-service)]
      servers:
        - url: http://localhost:9000
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/RegisterRequest"}
      responses:
        "200":
          description: >-
            Always 200. On success — {success:true, user}; on failure —
            {success:false, message:"email_already_registered"}. Sets the
            booknest_jwt/booknest_refresh cookies.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/AuthResponse"}
  /api/auth/login:
    post:
      summary: Log in (email/password)
      tags: [Auth (auth-service)]
      servers:
        - url: http://localhost:9000
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/LoginRequest"}
      responses:
        "200":
          description: >-
            Always 200. On failure — {success:false, message:"invalid_credentials"}.
            Sets the booknest_jwt/booknest_refresh cookies.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/AuthResponse"}
...
  /api/auth/me:
    get:
      summary: Current user
      tags: [Auth (auth-service)]
      servers:
        - url: http://localhost:9000
      security: [{jwtCookieAuth: []}]
      responses:
        "200":
          description: >-
            {success:false, message:"not_authenticated"|"account_not_found"}
            on failure, otherwise {success:true, user}.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/AuthResponse"}
    put:
      summary: Update the name and/or password
      tags: [Auth (auth-service)]
      servers:
        - url: http://localhost:9000
      security: [{jwtCookieAuth: []}]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/UpdateProfileRequest"}
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema: {$ref: "#/components/schemas/AuthResponse"}
  /api/auth/forgot-password:
    post:
      summary: Request a password recovery code
      description: >-
        SMTP is not configured (local/educational deployment) — the code is
        returned directly in the response (`devCode`); no emails are sent.
      tags: [Auth (auth-service)]
      servers:
        - url: http://localhost:9000
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/ForgotPasswordRequest"}
      responses:
        "200":
          description: >-
            {success:false, message:"account_not_found"} on failure,
            otherwise {success:true, email, sent:false, devCode}.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/ForgotPasswordResponse"}
  /api/auth/reset-password:
    post:
      summary: Reset the password using a code
      tags: [Auth (auth-service)]
      servers:
        - url: http://localhost:9000
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/ResetPasswordRequest"}
      responses:
        "200":
          description: >-
            {success:false, message:"invalid_or_expired_code"} on failure,
            otherwise {success:true, user}. Sets the
            booknest_jwt/booknest_refresh cookies.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/AuthResponse"}
...
components:
  securitySchemes:
    jwtCookieAuth:
      type: apiKey
      in: cookie
      name: booknest_jwt
      description: >-
        A JWT (HS256) issued by the Spring Boot auth-service after
        login/registration/Google OAuth. Rails only verifies the signature
        and the `role` field — it does not store its own credentials. The
        same cookie is also checked by auth-service itself
        (`/api/auth/me`, `PUT /api/auth/me`).
  schemas:
...
    RegisterRequest:
      type: object
      required: [email, password, name]
      properties:
        email: {type: string, format: email}
        password: {type: string, format: password, minLength: 6, maxLength: 128}
        name: {type: string, minLength: 1, maxLength: 100}
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email: {type: string, format: email}
        password: {type: string, format: password}
    UpdateProfileRequest:
      type: object
      description: Both fields are optional — only the ones provided are applied.
      properties:
        name: {type: string, minLength: 1, maxLength: 100}
        password: {type: string, format: password, minLength: 6, maxLength: 128}
    ForgotPasswordRequest:
      type: object
      required: [email]
      properties:
        email: {type: string, format: email}
    ForgotPasswordResponse:
      type: object
      properties:
        success: {type: boolean}
        email: {type: string, format: email, nullable: true}
        sent:
          {
            type: boolean,
            nullable: true,
            description: "Always false — SMTP is not configured",
          }
        devCode:
          {
            type: string,
            nullable: true,
            description: "Recovery code (instead of an email)",
          }
        message:
          {
            type: string,
            nullable: true,
            description: "Present only when success:false",
          }
    ResetPasswordRequest:
      type: object
      required: [email, code, newPassword]
      properties:
        email: {type: string, format: email}
        code: {type: string}
        newPassword:
          {type: string, format: password, minLength: 6, maxLength: 128}
    UserDto:
      type: object
      properties:
        username:
          {
            type: string,
            format: email,
            description: "Email — the primary identifier",
          }
        nickname: {type: string, description: "Display name"}
        role: {type: string, enum: [user, admin]}
    AuthResponse:
      type: object
      properties:
        success: {type: boolean}
        user: {allOf: [{$ref: "#/components/schemas/UserDto"}], nullable: true}
        message:
          {
            type: string,
            nullable: true,
            description: "Error code, present only when success:false",
          }